requirements.txt
beautifulsoup4==4.6.3
boto==2.49.0
boto3==1.9.18
botocore==1.12.18
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Django==2.1.2
django-admin-sortable==2.1.8
django-appconf==1.0.2
django-debug-toolbar==1.10.1
django-htmlmin==0.10.0
django-jet==1.0.8
django-parler==1.9.2
django-polymorphic==2.0.3
django-ratelimit==1.1.0
-e git+https://github.com/jazzband/django-robots.git@f484c2a6abcae6244f860ad077af28d4be62037e#egg=django_robots
django-storages==1.7.1
docutils==0.14
feedparser==5.2.1
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.12.0
idna==2.7
jmespath==0.9.3
mistune==0.8.3
mistune-contrib==0.1
oauth2client==4.1.3
Pillow==5.3.0
psycopg2==2.7.5
pur==5.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
requests==2.20.1
rsa==4.0
s3transfer==0.1.13
sentry-sdk==0.6.5
six==1.11.0
sorl-thumbnail==12.5.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.23
webencodings==0.5.1
